requirements.txt
pytube==9.3.5
python-slugify==1.2.6
gspread==3.1.0
oauth2client==4.1.3
pyOpenSSL==18.0.0
progress==1.4
